requirements.txt
Flask==0.10.1
PyJWT==1.4.0
# Application server for both development and production.
gunicorn==19.4.5

# Testing and static analysis.
pytest==2.9.1
pytest-cov==2.2.1
mock==1.3.0
flake8==2.5.4

# CLI.
Click==6.4

# Data and workers.
redis
celery

Flask-And-Redis
requests
flask_inputs
jsonschema
PyDrive
Flask-Babel==0.9
slackclient==1.0.0
Flask-Cache
cloudinary
kombu
gevent

